third_party/WebKit/LayoutTests/css3/flexbox/bug750553.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/check-layout-th.js"></script>
<style>
html, body {
  margin: 0;
}

body {
  display: flex;
  flex-direction: column;
}

.content {
  align-self: center;
}

.content > div {
  width: 400px;
  display: inline-block;
}
</style>

<body onload="checkLayout('.content')">
<div class="content" data-offset-x="0">
  <div data-offset-x="0">X</div>
  <div>X</div>
  <div>X</div>
  <div>X</div>
  <div>X</div>
</div>
